feat: migrate index to /app
