--- app/appDirRoute/page.tsx
@@ -1,5 +1,0 @@
-import Container from "@/components/Container";
-
-export default async function Page() {
-  return <Container> Hej from appdir</Container>;
-}
